chore(gateway): move official runner into gateway
