+++ gateway/runner/platform/macos.cjs
@@ -1,0 +1,154 @@
+const fs = require("fs");
+const path = require("path");
+const { execFileSync } = require("child_process");
+const {
+  RUNNER_APP_NAME,
+  RUNNER_BUNDLE_IDENTIFIER,
+  RUNNER_EXECUTABLE_NAME,
+} = require("../shared/constants.cjs");
+const {
+  fileFingerprint,
+  isDirectory,
+  realpathSafe,
+  readJsonIfPresent,
+  writeJson,
+} = require("../shared/fs-utils.cjs");
+const { logLine } = require("../shared/logging.cjs");
+const { writeGatewayAsar } = require("../shared/runner-asar.cjs");
+
+function escapePlistString(value) {
+  return String(value || "")
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;");
+}
+
+function runnerInfoPlist() {
+  /**
+   * runner 是后台 gateway 进程，不应该在 Dock 里出现第二个图标。
+   * LSUIElement 会让当前官方 Electron runtime 在 ChromeMain 阶段 SIGTRAP；
+   * LSBackgroundOnly 可以让进程成为真正的后台进程，同时仍能创建隐藏 webContents 承接官方 IPC。
+   */
+  return `<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+  <key>CFBundleExecutable</key>
+  <string>${escapePlistString(RUNNER_EXECUTABLE_NAME)}</string>
+  <key>CFBundleIdentifier</key>
+  <string>${escapePlistString(RUNNER_BUNDLE_IDENTIFIER)}</string>
+  <key>CFBundleName</key>
+  <string>OpenCodex Gateway</string>
+  <key>CFBundlePackageType</key>
+  <string>APPL</string>
+  <key>LSMinimumSystemVersion</key>
+  <string>12.0</string>
+  <key>NSHighResolutionCapable</key>
+  <true/>
+  <key>NSPrincipalClass</key>
+  <string>AtomApplication</string>
+  <key>LSBackgroundOnly</key>
+  <true/>
+</dict>
+</plist>
+`;
+}
+
+function officialFrameworksFingerprint(layout) {
+  const entries = [];
+  try {
+    for (const entry of fs.readdirSync(layout.frameworksDir, { withFileTypes: true })) {
+      entries.push({
+        name: entry.name,
+        kind: entry.isDirectory() ? "dir" : entry.isFile() ? "file" : entry.isSymbolicLink() ? "symlink" : "other",
+        stat: fileFingerprint(path.join(layout.frameworksDir, entry.name)),
+      });
+    }
+  } catch {}
+  return {
+    source: realpathSafe(layout.frameworksDir),
+    appRoot: realpathSafe(layout.appRoot),
+    executable: fileFingerprint(layout.executablePath),
+    asar: fileFingerprint(layout.asarPath),
+    entries,
+  };
+}
+
+function sameFrameworksFingerprint(left, right) {
+  return JSON.stringify(left || null) === JSON.stringify(right || null);
+}
+
+function ensureFrameworksCopy({ layout, runnerFrameworksDir, markerPath, logger }) {
+  const nextFingerprint = officialFrameworksFingerprint(layout);
+  const previous = readJsonIfPresent(markerPath);
+  if (
+    previous &&
+    sameFrameworksFingerprint(previous.fingerprint, nextFingerprint) &&
+    isDirectory(runnerFrameworksDir)
+  ) {
+    logLine(logger, `official Electron Frameworks cache hit: ${runnerFrameworksDir}`);
+    return { copied: false };
+  }
+
+  /**
+   * 不能把 Contents/Frameworks 做成指向 /Applications 的符号链接：官方 helper 进程会按 runner
+   * 内部路径回溯加载 framework，macOS sandbox 会拒绝跨 bundle symlink。这里复制到运行态 cache。
+   */
+  fs.rmSync(runnerFrameworksDir, { recursive: true, force: true });
+  fs.cpSync(layout.frameworksDir, runnerFrameworksDir, { recursive: true, force: true });
+  writeJson(markerPath, {
+    fingerprint: nextFingerprint,
+    copiedAt: new Date().toISOString(),
+  });
+  logLine(logger, `official Electron Frameworks copied: ${layout.frameworksDir} -> ${runnerFrameworksDir}`);
+  return { copied: true };
+}
+
+function signRunnerExecutable(executablePath) {
+  /**
+   * 官方主二进制复制到 runner 后原签名会失效；只重签这个入口壳。
+   * Frameworks 保留官方签名副本，避免 deep 签名破坏 Chromium framework 的封装结构。
+   */
+  execFileSync("/usr/bin/codesign", ["--force", "--sign", "-", executablePath], {
+    stdio: ["ignore", "ignore", "pipe"],
+  });
+}
+
+async function createMacRunner({ layout, runtimeDir, logger }) {
+  const workDir = path.join(runtimeDir, "official-electron-runner");
+  const runnerAppPath = path.join(workDir, RUNNER_APP_NAME);
+  const contentsDir = path.join(runnerAppPath, "Contents");
+  const runnerMacosDir = path.join(contentsDir, "MacOS");
+  const runnerResourcesDir = path.join(contentsDir, "Resources");
+  const runnerFrameworksDir = path.join(contentsDir, "Frameworks");
+  const frameworksMarkerPath = path.join(workDir, "frameworks-manifest.json");
+  const runnerExecutablePath = path.join(runnerMacosDir, RUNNER_EXECUTABLE_NAME);
+
+  // 官方升级后 framework/ABI 可能变化，Frameworks 由 fingerprint 控制；入口和 app.asar 每次重写。
+  fs.rmSync(runnerMacosDir, { recursive: true, force: true });
+  fs.rmSync(runnerResourcesDir, { recursive: true, force: true });
+  fs.mkdirSync(runnerMacosDir, { recursive: true });
+  fs.mkdirSync(runnerResourcesDir, { recursive: true });
+
+  fs.copyFileSync(layout.executablePath, runnerExecutablePath);
+  fs.chmodSync(runnerExecutablePath, 0o755);
+  ensureFrameworksCopy({ layout, runnerFrameworksDir, markerPath: frameworksMarkerPath, logger });
+  fs.writeFileSync(path.join(contentsDir, "Info.plist"), runnerInfoPlist(), "utf8");
+  fs.writeFileSync(path.join(contentsDir, "PkgInfo"), "APPL????", "utf8");
+  await writeGatewayAsar({ runnerResourcesDir, workDir });
+  signRunnerExecutable(runnerExecutablePath);
+
+  logLine(logger, `prepared official Electron runner: app=${runnerAppPath}`);
+  logLine(logger, `official Electron source: app=${layout.appRoot} asar=${layout.asarPath}`);
+
+  return {
+    executablePath: runnerExecutablePath,
+    runnerAppPath,
+    officialAppPath: layout.appRoot,
+    officialAsarPath: layout.asarPath,
+  };
+}
+
+module.exports = {
+  createMacRunner,
+};
