+++ gateway/runner/shared/constants.cjs
@@ -1,0 +1,13 @@
+const RUNNER_APP_NAME = "OpenCodex Gateway.app";
+const RUNNER_EXECUTABLE_NAME = "Codex";
+const RUNNER_WINDOWS_EXECUTABLE_NAME = "OpenCodex Gateway.exe";
+const RUNNER_LINUX_EXECUTABLE_NAME = "opencodex-gateway";
+const RUNNER_BUNDLE_IDENTIFIER = "dev.opencodex.gateway.officialruntime";
+
+module.exports = {
+  RUNNER_APP_NAME,
+  RUNNER_EXECUTABLE_NAME,
+  RUNNER_WINDOWS_EXECUTABLE_NAME,
+  RUNNER_LINUX_EXECUTABLE_NAME,
+  RUNNER_BUNDLE_IDENTIFIER,
+};
